+++ app/src/main/java/com/osipenko/pomodoro/domain/TaskItem.kt
@@ -2,9 +2,9 @@ package com.osipenko.pomodoro.domain
 
 data class TaskItem (
     val text: String,
-    var id: Int = UNDEFINED_ID
+    var id: Long = UNDEFINED_ID
 ) {
     companion object {
-        const val UNDEFINED_ID = 0
+        const val UNDEFINED_ID = 0L
     }
 }
